+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # ChangeLog
 
+## [v0.0.3] - 2024-12-13
+
+- ADD: use `A2A's` write code service, instead of raw OpenAI's API
+- DEL: remove `openai` dependency and 'useLlm' service
+
 ## [v0.0.2] - 2024-12-13
 
 - ADD: implement run function
-
 
 ## [v0.0.1] - 2024-11-19